Add toolchain librarypaths to -Wl,-rpath (#1934)
* add lib/lib64 toolchain rpaths for execution

* refactor to not have duplicate code
