+++ lib/base-compiler.js
@@ -36,7 +36,8 @@ const temp = require('temp'),
     argumentParsers = require("./compilers/argument-parsers"),
     CompilerArguments = require("./compiler-arguments"),
     cfg = require('./cfg'),
-    languages = require('./languages').list;
+    languages = require('./languages').list,
+    toolchainUtils = require('./toolchain-utils');
 
 class BaseCompiler {
     constructor(compilerInfo, env) {
@@ -63,6 +64,8 @@ class BaseCompiler {
 
         this.asm = new AsmParser(this.compilerProps);
         this.llvmIr = new LlvmIrParser(this.compilerProps);
+
+        this.toolchainPath = toolchainUtils.getToolchainPath(this.compiler.exe, this.compiler.options);
 
         this.possibleArguments = new CompilerArguments(this.compiler.id);
         this.possibleTools = _.values(compilerInfo.tools);
@@ -332,8 +335,17 @@ class BaseCompiler {
         const pathFlag = this.compiler.rpathFlag || "-Wl,-rpath,";
         const libPathFlag = this.compiler.libpathFlag || "-L";
 
+        let toolchainLibraryPaths = [];
+        if (this.toolchainPath) {
+            toolchainLibraryPaths = [
+                path.join(this.toolchainPath, '/lib64'),
+                path.join(this.toolchainPath, '/lib32')
+            ];
+        }
+
         return _.union(
             this.compiler.ldPath.map(path => pathFlag + path),
+            toolchainLibraryPaths.map(path => pathFlag + path),
             this.getSharedLibraryPaths(libraries).map(path => pathFlag + path),
             this.getSharedLibraryPaths(libraries).map(path => libPathFlag + path));
     }
