+++ lib/tooling/compiler-dropin-tool.js
@@ -26,7 +26,7 @@
 const
     BaseTool = require('./base-tool'),
     _ = require('underscore'),
-    path = require('path');
+    toolchainUtils = require('../toolchain-utils');
 
 class CompilerDropinTool extends BaseTool {
     constructor(toolInfo, env) {
@@ -34,21 +34,7 @@ class CompilerDropinTool extends BaseTool {
     }
 
     getToolchainPath(compilationInfo) {
-        const options = compilationInfo.compiler.options ? compilationInfo.compiler.options.split(" ") : [];
-        const existingChain = _.find(options, (elem) => elem.indexOf("--gcc-toolchain=") !== -1);
-
-        if (!existingChain) {
-            const gxxname = _.find(options, (elem) => elem.indexOf("--gxx-name=") !== -1);
-            if (gxxname) {
-                return path.resolve(path.dirname(gxxname.substr(11)), "..");
-            } else if (compilationInfo.compiler.exe.indexOf("/g++") !== -1) {
-                return path.resolve(path.dirname(compilationInfo.compiler.exe), "..");
-            } else {
-                return false;
-            }
-        } else {
-            return existingChain.substr(16);
-        }
+        return toolchainUtils.getToolchainPath(compilationInfo.compiler.exe, compilationInfo.compiler.options);
     }
 
     getOrderedArguments(compilationInfo, includeflags, args, sourcefile) {
